requirements.txt
cv2
numpy
tkinter
pillow